+++ src/Concerto/PanelBundle/Resources/starter_content/src/template/itemGracelyScale/itemGracelyScale.js
@@ -20,6 +20,10 @@ testRunner.compileProvider.component('itemGracelyScale', {
       $scope.item = this.item;
       $scope.response = this.response;
       $scope.responseRequired = this.responseRequired;
+      
+      $scope.response.isValid = function() {
+        return $scope.intensity != null || $scope.unpleasantness != null;
+      }
 
       $scope.unpleasantnessVisible = typeof($scope.item.responseOptions.gracelyScaleShow) === 'undefined' || $scope.item.responseOptions.gracelyScaleShow === 'both' || $scope.item.responseOptions.gracelyScaleShow === 'unpleasantness';
       $scope.intensityVisible = typeof($scope.item.responseOptions.gracelyScaleShow) === 'undefined' || $scope.item.responseOptions.gracelyScaleShow === 'both' || $scope.item.responseOptions.gracelyScaleShow === 'intensity';