+++ src/Concerto/PanelBundle/Resources/starter_content/src/template/itemOptions/itemOptions.js
@@ -3,7 +3,7 @@ testRunner.compileProvider.component('itemOptions', {
   bindings: {
     item: '=',
     response: '=',
-    responseRequired: "<"
+    responseRequired: '<'
   },
   controller: function controller($scope) {
     this.$onInit = function() {