+++ src/Concerto/PanelBundle/Resources/starter_content/src/template/test/test.js
@@ -11,9 +11,22 @@ testRunner.controllerProvider.register("test", function ($scope) {
   $scope.backButtonLabel = testRunner.R.backButtonLabel;
   $scope.showPageInfo = testRunner.R.showPageInfo == 1;
 
+  $scope.isFormValid = function() {
+    if(!$scope.responseRequired) {
+      return true;
+    }
+    for(let i=0; i<$scope.items.length; i++) {
+      let item = $scope.items[i];
+      let response = $scope.responses["r" + item.id];
+      if(response.skipped) continue;
+      if(!response.isValid()) return false;
+    }
+    return true;
+  }
+
   if ($scope.pastResponses) {
-    for (var i = 0; i < $scope.pastResponses.length; i++) {
-      var response = $scope.pastResponses[i];
+    for (let i = 0; i < $scope.pastResponses.length; i++) {
+      let response = $scope.pastResponses[i];
       $scope.responses["r" + response.item_id] = typeof response.response === 'object' ? response.response : {
         value: response.response,
         skipped: response.skipped
@@ -21,19 +34,21 @@ testRunner.controllerProvider.register("test", function ($scope) {
     }
   }
 
-  for (var i = 0; i < $scope.items.length; i++) {
-    var item = $scope.items[i];
+  for (let i = 0; i < $scope.items.length; i++) {
+    let item = $scope.items[i];
     if (typeof item.responseOptions === "string") {
       item.responseOptions = JSON.parse(item.responseOptions);
-      $scope.items[i] = item;
     }
-    $scope.items[i] = item;
-
     if (typeof $scope.responses["r" + item.id] === 'undefined') {
       $scope.responses["r" + item.id] = {
         skipped: 0
       };
     }
+    
+    $scope.responses["r" + item.id].isValid = function() {
+      return this.value;
+    }
+    
     testRunner.addExtraControl("skip"+item.id, function() {
       return $scope.responses["r"+item.id].skipped;
     });